Refactored 'search' command (#80)
